+++ docs/CHANGELOG.md
@@ -57,6 +57,7 @@
     * `Text Rendering Fixes` (Enabled By Default)
     * `Close Editor When Sign Is Destroyed` (Enabled By Default)
     * `Remove Chest Placement Restrictions` (Enabled By Default)
+    * `Fix Hanging When No Valid Spawn Point Exists` (Enabled By Default)
   * Existing Functionality (All Enabled By Default)
     * `Fix Screen Rendering When Hiding HUD`
     * `Sanitize Usernames`